pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
description = ""
authors = ["Micheal Lanham <cxbxmxcx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.17.1"
streamlit = "^1.32.0"
openai = "^1.13.3"
streamlit-js-eval = "^0.1.7"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
wikipedia = "^1.4.0"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
anthropic = "^0.20.0"
groq = "^0.4.2"
langchain = "^0.1.13"
chromadb = "^0.4.24"
plotly = "^5.20.0"
scikit-learn = "^1.4.1.post1"
barfi = "^0.7.0"
streamlit-code-editor = "^0.1.12"
lark-parser = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
nexus = "ai_agent_nexus.cli:main"


